fix estructuracion de archivos
